Finished axios requests
Import "RequestUserData" (file is in src/features) to get functions to add/get a single Journal of a specified date or a list of habits. It uses POST requests as they are simpler to use.

The "requestPerDates" function in the "archive" requests journal entries per date given a start and end date. It can also be used for the "weeklyView".

All axios requests are done in the features folder.

"LogIn" mock up is placed in routes successfully inserts java token to localStorage. The java token is saved to the key "user".
"Logout" is complete and deletes token from localStorage
"Sign up" is complete and saves token to localStorage

install "jsonwebtoken" and  "axios" to run program on your computer

"authMiddleware.js" has a function "isValidToken" that returns a username if a given token is valid, else it returns null. This function already confirms a correct password exists for the username.